commit with rename of generator
