+++ src/main/resources/test.json
@@ -1,36 +1,40 @@
 {
   "detectives": [
     {
-       "MainId": 1,
-       "firstName": "Sherlock",
-       "lastName": "Homes",
-       "violinPlayer": true,
-       "categories":         [
-            {
-                 "CategoryID": 1,
-                 "CategoryName": "extras1",
-                 "extra": {
-                   "extraArray": [
-                         {"violin":1},
-                         {"cap":2}
-                        ]
-                     }
-                 }
-             ]
-        },
+      "MainId": 1,
+      "firstName": "Sherlock",
+      "lastName": "Homes",
+      "violinPlayer": true,
+      "categories": [
         {
-                "MainId": 3,
-                "firstName": "James",
-                "lastName": "Watson",
-                "violinPlayer": false,
-                "categories": [
-                                  {
-                                      "CategoryID": 2,
-                                      "CategoryName": "extras2",
-                                      "extra": null
-                                  }
-                             ]
-            }
-    ],
-    "success": true
+          "CategoryID": 1,
+          "CategoryName": "extras1",
+          "extra": {
+            "extraArray": [
+              {
+                "violin": 1
+              },
+              {
+                "cap": 2
+              }
+            ]
+          }
+        }
+      ]
+    },
+    {
+      "MainId": 3,
+      "firstName": "James",
+      "lastName": "Watson",
+      "violinPlayer": false,
+      "categories": [
+        {
+          "CategoryID": 2,
+          "CategoryName": "extras2",
+          "extra": null
+        }
+      ]
+    }
+  ],
+  "success": true
 }
